requirements_new.txt
# Core dependencies
fastapi==0.128.1
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# Database
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11

# LLM providers
openai==2.17.0
ollama==0.6.1

# Semantic Search & Embeddings (BGE-M3)
sentence-transformers==3.3.1
torch>=2.0.0
numpy>=1.24.0

# Utilities
tqdm==4.67.3
colorama==0.4.6
pytz==2025.2

# HTTP clients
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
anyio==4.12.1
certifi==2026.1.4
idna==3.11
sniffio==1.3.1

# Other dependencies
annotated-doc==0.0.4
annotated-types==0.7.0
click==8.3.1
distro==1.9.0
greenlet==3.3.1
jiter==0.13.0
pydantic_core==2.41.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0

